i2c: meson: add suspend/resume to set gpio status [1/1]
PD#SWPL-17544

Problem:
it will cause Current reverse irrigation if the gpio
to i2c function when the system is in STR mode

Solution:
keep the gpio to sleep mode to protect to slave device

there are two sleep mode for i2c pins, and it depends the
pull up power. we can choose the exact sleep mode.

1)pull up power is on, set the sleep mode to input
2)pull up power if off, set low the gpio

Verify:
verified on x301

Change-Id: Ie62f70449be2ff2cef7e3ab5c918ac379d861ca1
Signed-off-by: Jian Hu <jian.hu@amlogic.com>