avsync: tsync: fix the dtv avsync problem [1/2]
PD#SWPL-16491

Problem:
PD#SWPL-16491, PD#SWPL-17385
avsync is lost sometimes in some special strems.

Solution:
use the vpts as the system clock if the pcr and apts
are bigger than the vpts when pts checkin.
add the first audio pts offset for audiohal when
the audio pts is looked up

Verify:
X301

Change-Id: I7f4a953322c390db7086188db9bcb286fc55b0ed
Signed-off-by: Xiaoming Sui <xiaoming.sui@amlogic.com>