tsync: fix tsync init error after str resume [1/1]
PD#SWPL-19936

Problem:
the tsync pcr was init after str resume but no signal.
in this case, it is an invalid status, it should be inited
after audio and/or video data are resumed.

Solution:
update first_time_record time if pcr and audio/video
checkin pts are not inited.

Verify:
X301

Change-Id: Iefe2d57b738599f1fac9a66825d3f4d6c823e545
Signed-off-by: Xiaoming Sui <xiaoming.sui@amlogic.com>